Update NodeRpc and rearrange structure
