+++ altchains/test/.gitignore
@@ -1,0 +1,1 @@
+!__init__.py
