+++ altchains/test/load_test/__init__.py
@@ -1,8 +1,3 @@
-import time
-
-from rpc.random_miner import RandomMiner
-
-
 def load_test(node, apm, max_blocks, max_hours, seed=None):
     info = node.getblockchaininfo()
     chain = info['chain']
@@ -12,17 +7,16 @@ def load_test(node, apm, max_blocks, max_hours, seed=None):
     if blocks > 0:
         raise Exception('Unexpected blocks count (expected: 0, actual: {})'.format(blocks))
 
-    miner = RandomMiner(node, apm, seed)
+    from .random_miner import RandomMiner
+    from time import time
 
-    start = time.time()
+    miner = RandomMiner(node, apm, seed)
+    start = time()
 
     while True:
         blocks = node.getblockcount()
-        elapsed = time.time() - start
+        elapsed = time() - start
         if blocks >= max_blocks or elapsed >= max_hours * 60 * 60:
             return blocks, elapsed
 
         miner.execute_command()
-
-
-
